Remove receipt_index_from_proof
