+++ bridges/snowbridge/pallets/ethereum-client/fixtures/Cargo.toml
@@ -32,4 +32,4 @@ std = [
 	"sp-core/std",
 	"sp-std/std",
 ]
-runtime-benchmarks = ["snowbridge-verification-primitives/runtime-benchmarks"]
+runtime-benchmarks = []
