+++ bridges/snowbridge/pallets/ethereum-client/fixtures/src/lib.rs
@@ -13,9 +13,6 @@ use snowbridge_verification_primitives::{EventFixture, EventProof, Log, Proof};
 use sp_core::U256;
 use sp_std::{boxed::Box, vec};
 
-#[cfg(any(feature = "runtime-benchmarks", feature = "std", test))]
-use snowbridge_verification_primitives::receipt_helper::receipt_index_from_proof;
-
 const SC_SIZE: usize = 512;
 const SC_BITS_SIZE: usize = 64;
 type CheckpointUpdate = snowbridge_beacon_primitives::CheckpointUpdate<SC_SIZE>;
@@ -1240,8 +1237,7 @@ pub fn make_inbound_fixture() -> EventFixture {
         hex!("f851a0925fd57a02e48ba0efa2dc468e694b47f3f7b5443b944b4b2ef05c24deb247e280808080808080a0c0127fe02b1e3c9608db2985cbbc8cecc64fb6ce4477e1071d61bfd127a4027b8080808080808080").to_vec(),
         hex!("f9034f30b9034bf90348018301789ab9010000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000810000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000001000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000f9023df9023a94b1185ede04202fe62d38f5db72f71e38ff3e8305e1a0550e2067494b1736ea5573f2d19cdc0ac95b410fff161bf16f11c6229655ec9cb9020000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000040000000000000000000000000b1185ede04202fe62d38f5db72f71e38ff3e830500000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001a0000000000000000000000000000000000000000000000000000009184e72a0000000000000000000000000000000000000000000000000000000015d3ef798000000000000000000000000000000000000000000000000000000015d3ef798000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000040000000000000000000000000b8ea8cb425d85536b158d661da1ef0895bb92f1d00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000").to_vec(),
     ];
-	let receipt_index = receipt_index_from_proof(&receipt_proof)
-		.expect("receipt index should be derivable from proof");
+	let tx_index = 0u64;
 	EventFixture {
         event: EventProof {
             event_log: 	Log {
@@ -1250,9 +1246,9 @@ pub fn make_inbound_fixture() -> EventFixture {
                     hex!("550e2067494b1736ea5573f2d19cdc0ac95b410fff161bf16f11c6229655ec9c").into(),
                 ],
                 data: hex!("00000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000040000000000000000000000000b1185ede04202fe62d38f5db72f71e38ff3e830500000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001a0000000000000000000000000000000000000000000000000000009184e72a0000000000000000000000000000000000000000000000000000000015d3ef798000000000000000000000000000000000000000000000000000000015d3ef798000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000040000000000000000000000000b8ea8cb425d85536b158d661da1ef0895bb92f1d00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000").into(),
+                tx_index,
             },
             proof: Proof {
-                receipt_index,
                 receipt_proof,
                 execution_proof: ExecutionProof {
                     header: BeaconHeader {
