+++ bridges/snowbridge/pallets/inbound-queue-v2/fixtures/src/lib.rs
@@ -2,5 +2,4 @@
 // SPDX-FileCopyrightText: 2023 Snowfork <hello@snowfork.com>
 #![cfg_attr(not(feature = "std"), no_std)]
 
-#[cfg(any(feature = "runtime-benchmarks", feature = "std", test))]
 pub mod register_token;
