+++ bridges/snowbridge/pallets/inbound-queue-v2/src/test.rs
@@ -25,7 +25,6 @@ fn test_submit_happy_path() {
 		let event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -57,7 +56,6 @@ fn test_submit_with_invalid_gateway() {
 		let event = EventProof {
 			event_log: mock_event_log_invalid_gateway(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -80,7 +78,6 @@ fn test_submit_verification_fails_with_invalid_proof() {
 		let mut event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -106,7 +103,6 @@ fn test_submit_fails_with_malformed_message() {
 		let event = EventProof {
 			event_log: mock_event_log_invalid_message(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -130,7 +126,6 @@ fn test_using_same_nonce_fails() {
 		let event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -163,7 +158,6 @@ fn test_set_operating_mode() {
 		let event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -215,7 +209,6 @@ fn test_xcm_send_failure() {
 		let event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -246,7 +239,6 @@ fn test_xcm_send_validate_failure() {
 		let event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -272,7 +264,6 @@ fn test_xcm_charge_fees_failure() {
 		let event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -293,7 +284,6 @@ fn test_register_token() {
 		let event = EventProof {
 			event_log: mock_event_log_v2(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -311,7 +301,6 @@ fn test_switch_operating_mode() {
 		let event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
