+++ bridges/snowbridge/pallets/inbound-queue/fixtures/src/lib.rs
@@ -2,11 +2,7 @@
 // SPDX-FileCopyrightText: 2023 Snowfork <hello@snowfork.com>
 #![cfg_attr(not(feature = "std"), no_std)]
 
-#[cfg(any(feature = "runtime-benchmarks", feature = "std", test))]
 pub mod register_token;
-#[cfg(any(feature = "runtime-benchmarks", feature = "std", test))]
 pub mod send_native_eth;
-#[cfg(any(feature = "runtime-benchmarks", feature = "std", test))]
 pub mod send_token;
-#[cfg(any(feature = "runtime-benchmarks", feature = "std", test))]
 pub mod send_token_to_penpal;
