+++ bridges/snowbridge/pallets/inbound-queue/src/test.rs
@@ -26,7 +26,6 @@ fn test_submit_happy_path() {
 		let event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -77,7 +76,6 @@ fn test_submit_xcm_invalid_channel() {
 		let event = EventProof {
 			event_log: mock_event_log_invalid_channel(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -103,7 +101,6 @@ fn test_submit_with_invalid_gateway() {
 		let event = EventProof {
 			event_log: mock_event_log_invalid_gateway(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -129,7 +126,6 @@ fn test_submit_with_invalid_nonce() {
 		let event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -163,7 +159,6 @@ fn test_submit_no_funds_to_reward_relayers_just_ignore() {
 		let event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -181,7 +176,6 @@ fn test_set_operating_mode() {
 		let event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -223,7 +217,6 @@ fn test_submit_no_funds_to_reward_relayers_and_ed_preserved() {
 		let event = EventProof {
 			event_log: mock_event_log(),
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
@@ -240,7 +233,6 @@ fn test_submit_no_funds_to_reward_relayers_and_ed_preserved() {
 		let event = EventProof {
 			event_log,
 			proof: Proof {
-				receipt_index: 0,
 				receipt_proof: Default::default(),
 				execution_proof: mock_execution_proof(),
 			},
