+++ bridges/snowbridge/pallets/outbound-queue-v2/Cargo.toml
@@ -85,7 +85,6 @@ runtime-benchmarks = [
 	"pallet-message-queue/runtime-benchmarks",
 	"snowbridge-core/runtime-benchmarks",
 	"snowbridge-test-utils/runtime-benchmarks",
-	"snowbridge-verification-primitives/runtime-benchmarks",
 	"sp-runtime/runtime-benchmarks",
 	"xcm-builder/runtime-benchmarks",
 	"xcm-executor/runtime-benchmarks",
