+++ bridges/snowbridge/pallets/outbound-queue-v2/src/fixture.rs
@@ -5,21 +5,18 @@
 
 use hex_literal::hex;
 use snowbridge_beacon_primitives::{
-    types::deneb, AncestryProof, BeaconHeader, ExecutionProof, VersionedExecutionPayloadHeader,
+	types::deneb, AncestryProof, BeaconHeader, ExecutionProof, VersionedExecutionPayloadHeader,
 };
-use snowbridge_verification_primitives::{
-    receipt_helper::receipt_index_from_proof, EventFixture, EventProof, Log, Proof,
-};
+use snowbridge_verification_primitives::{EventFixture, EventProof, Log, Proof};
 use sp_core::U256;
 use sp_std::vec;
 
 pub fn make_submit_delivery_receipt_message() -> EventFixture {
-    let receipt_proof = vec![
+	let receipt_proof = vec![
         hex!("f9024e822080b9024802f9024401830d716eb9010000200000000000000000000000000000080000000000000000000000000000000000000000000004000000000000000000000000000000000000000000008000000000000000000801000000000000000000000000000000000000000000000000000000020000000008000000000800000000020000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000800000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000020000000000200000000000000000000000000000000000000000000000000000000f90139f87a94b1185ede04202fe62d38f5db72f71e38ff3e8305f842a057f58171b8777633d03aff1e7408b96a3d910c93a7ce433a8cb7fb837dc306a6a09441dceeeffa7e032eedaccf9b7632e60e86711551a82ffbbb0dda8afd9e4ef7a0000000000000000000000000de45448ca2d57797c0bec0ee15a1e42334744219f8bb94b1185ede04202fe62d38f5db72f71e38ff3e8305f842a08856ab63954e6c2938803a4654fb704c8779757e7bfdbe94a578e341ec637a95a00000000000000000000000000000000000000000000000000000000000000000b860907b6ec7bf3f2496ef79238e0fb19e032bfe444c7ffe906bd340c6c4ffe8511f0000000000000000000000000000000000000000000000000000000000000001d43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d").to_vec(),
     ];
-    let receipt_index = receipt_index_from_proof(&receipt_proof)
-        .expect("receipt index should be derivable from proof");
-    EventFixture {
+	let tx_index = 0u64;
+	EventFixture {
         event: EventProof {
             event_log: 	Log {
                 address: hex!("b1185ede04202fe62d38f5db72f71e38ff3e8305").into(),
@@ -28,9 +25,9 @@ pub fn make_submit_delivery_receipt_message() -> EventFixture {
                     hex!("0000000000000000000000000000000000000000000000000000000000000000").into(),
                 ],
                 data: hex!("907b6ec7bf3f2496ef79238e0fb19e032bfe444c7ffe906bd340c6c4ffe8511f0000000000000000000000000000000000000000000000000000000000000001d43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d").into(),
+                tx_index,
             },
             proof: Proof {
-                receipt_index,
                 receipt_proof,
                 execution_proof: ExecutionProof {
                     header: BeaconHeader {
