--- bridges/snowbridge/primitives/ethereum/README.md
@@ -1,4 +1,0 @@
-# Ethereum Primitives
-
-Contains code necessary to decode RLP encoded data (like the Ethereum log), structs for Ethereum execution headers. The
-code in this crate relates to the Ethereum execution chain.
