+++ bridges/snowbridge/primitives/inbound-queue/Cargo.toml
@@ -70,7 +70,6 @@ runtime-benchmarks = [
 	"frame-system/runtime-benchmarks",
 	"snowbridge-core/runtime-benchmarks",
 	"snowbridge-test-utils/runtime-benchmarks",
-	"snowbridge-verification-primitives/runtime-benchmarks",
 	"sp-runtime/runtime-benchmarks",
 	"xcm-builder/runtime-benchmarks",
 	"xcm-executor/runtime-benchmarks",
