+++ bridges/snowbridge/primitives/verification/src/lib.rs
@@ -11,9 +11,6 @@ use sp_std::prelude::*;
 
 pub mod receipt;
 
-#[cfg(any(feature = "runtime-benchmarks", feature = "std", test))]
-pub mod receipt_helper;
-
 /// A trait for verifying inbound messages from Ethereum.
 pub trait Verifier {
 	fn verify(event: &Log, proof: &Proof) -> Result<(), VerificationError>;
@@ -49,13 +46,12 @@ pub struct Log {
 	pub address: H160,
 	pub topics: Vec<H256>,
 	pub data: Vec<u8>,
+	pub tx_index: u64,
 }
 
 /// Inclusion proof for a transaction receipt
 #[derive(Clone, Encode, Decode, DecodeWithMemTracking, PartialEq, Debug, TypeInfo)]
 pub struct Proof {
-	// Receipt index in the block (trie key)
-	pub receipt_index: u64,
 	// Proof values from receipts tree
 	pub receipt_proof: Vec<Vec<u8>>,
 	// Proof that an execution header was finalized by the beacon chain
